+++ NEWS.md
@@ -8,6 +8,7 @@
 
 - Made input checks more robust for required questions -- necessary for some custom input types (#32).
 - Updated CSS for matrix questions to ensure their width covers the entire question container. This was a problem for some users with Google Chrome and Safari.
+- Fixed issue resulting in required matrix questions with instructions not working.
 
 # shinysurveys 0.2.0